local checks passed

+++ src/main/resources/db/scripts/001_ddl_create_food_table.sql
@@ -1,5 +1,5 @@
 CREATE TABLE food (
                       id BIGSERIAL PRIMARY KEY,
                       title VARCHAR(255) NOT NULL,
-                      price NUMERIC(10, 2) NOT NULL
+                      price float NOT NULL
 );
